+++ src/main/resources/config/automation_SCPAPI_QA.properties
@@ -32,3 +32,5 @@ scp.create.permission.endpoint=scp/v1/permissions/create
 scp.update.permission.endpoint=/scp/v1/permissions/update
 scp.create.entitytype.endpoint=/scp/v1/entity-types/create
 scp.update.entitytype.endpoint=/scp/v1/entity-types/update
+scp.qa.permission.list=scp/v1/permissions/getPermissions
+
